Apply comments. First steps.
