--- .github/workflows/tests.yml
@@ -1,31 +1,0 @@
-name: Tests
-
-on: [push, pull_request]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-16.04
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Install Python 3
-      run: |
-        curl -L https://github.com/pyenv/pyenv-installer/raw/master/bin/pyenv-installer | bash
-        export PATH=$HOME/.pyenv/shims:$HOME/.pyenv/bin:$PATH
-        pyenv install 3.7.3
-        pyenv global 3.7.3
-        pyenv rehash
-    - name: Install dependencies
-      run: |
-        export PATH=$HOME/.pyenv/shims:$HOME/.pyenv/bin:$PATH
-        pip install -r requirements_examples.txt
-    - name: Run unittests
-      run: |
-        export PATH=$HOME/.pyenv/shims:$HOME/.pyenv/bin:$PATH
-        python -m unittest discover .
-    - name: Test install
-      run: |
-        export PATH=$HOME/.pyenv/shims:$HOME/.pyenv/bin:$PATH
-        python setup.py install
-
