+++ requirements.txt
@@ -1,0 +1,117 @@
+actionlib==1.12.0
+angles==1.9.12
+attrs==19.3.0
+bondpy==1.8.3
+camera-calibration==1.14.0
+camera-calibration-parsers==1.11.13
+catkin==0.7.20
+certifi==2019.11.28
+chardet==3.0.4
+controller-manager==0.17.0
+controller-manager-msgs==0.17.0
+cv-bridge==1.13.0
+diagnostic-analysis==1.9.3
+diagnostic-common-diagnostics==1.9.3
+diagnostic-updater==1.9.3
+dynamic-reconfigure==1.6.0
+gazebo-plugins==2.8.6
+gazebo-ros==2.8.6
+gencpp==0.6.2
+geneus==2.2.6
+genlisp==0.4.16
+genmsg==0.5.12
+gennodejs==2.0.1
+genpy==0.6.9
+idna==2.9
+image-geometry==1.13.0
+interactive-markers==1.11.4
+joint-state-publisher==1.12.14
+kdl-parser-py==1.13.1
+laser-geometry==1.6.4
+message-filters==1.14.3
+more-itertools==8.2.0
+numpy==1.18.2
+packaging==20.3
+pluggy==0.13.1
+py==1.8.1
+pyparsing==2.4.6
+pytest==5.4.1
+pytest-runner==5.2
+python-qt-binding==0.4.0
+qt-dotgraph==0.4.0
+qt-gui==0.4.0
+qt-gui-cpp==0.4.0
+qt-gui-py-common==0.4.0
+requests==2.23.0
+resource-retriever==1.12.6
+rosbag==1.14.3
+rosboost-cfg==1.14.7
+rosclean==1.14.7
+roscreate==1.14.7
+rosgraph==1.14.3
+roslaunch==1.14.3
+roslib==1.14.7
+roslint==0.11.2
+roslz4==1.14.3
+rosmake==1.14.7
+rosmaster==1.14.3
+rosmsg==1.14.3
+rosnode==1.14.3
+rosparam==1.14.3
+rospy==1.14.3
+rosservice==1.14.3
+rostest==1.14.3
+rostopic==1.14.3
+rosunit==1.14.7
+roswtf==1.14.3
+rqt-action==0.4.9
+rqt-bag==0.4.12
+rqt-bag-plugins==0.4.12
+rqt-console==0.4.9
+rqt-dep==0.4.9
+rqt-graph==0.4.11
+rqt-gui==0.5.0
+rqt-gui-py==0.5.0
+rqt-image-view==0.4.14
+rqt-launch==0.4.8
+rqt-logger-level==0.4.8
+rqt-moveit==0.5.7
+rqt-msg==0.4.8
+rqt-nav-view==0.5.7
+rqt-plot==0.4.9
+rqt-pose-view==0.5.8
+rqt-publisher==0.4.8
+rqt-py-common==0.5.0
+rqt-py-console==0.4.8
+rqt-reconfigure==0.5.1
+rqt-robot-dashboard==0.5.7
+rqt-robot-monitor==0.5.9
+rqt-robot-steering==0.5.10
+rqt-runtime-monitor==0.5.7
+rqt-rviz==0.6.0
+rqt-service-caller==0.4.8
+rqt-shell==0.4.9
+rqt-srv==0.4.8
+rqt-tf-tree==0.6.0
+rqt-top==0.4.8
+rqt-topic==0.4.11
+rqt-web==0.4.8
+rviz==1.13.7
+scipy==1.4.1
+sensor-msgs==1.12.7
+six==1.14.0
+smach==2.0.1
+smach-ros==2.0.1
+smclib==1.8.3
+tf==1.12.0
+tf-conversions==1.12.0
+tf2-geometry-msgs==0.6.5
+tf2-kdl==0.6.5
+tf2-py==0.6.5
+tf2-ros==0.6.5
+topic-tools==1.14.3
+tqdm==4.43.0
+urdfdom-py==0.4.2
+urllib3==1.25.8
+wcwidth==0.1.9
+xacro==1.13.5
